pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loom"
version = "1.2.7.dev20251120"
description = "Tailor resumes using the OpenAI Responses API"
requires-python = ">=3.12"
dependencies = [
    "typer",
    "python-docx",
    "openai",
    "anthropic",
    "ollama",
    "python-dotenv",
    "rich",
    "simple-term-menu",
    "readchar",
]

[project.scripts]
loom = "src.main:app"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-socket",
    "pytest-xdist",
    "freezegun",
    "faker",
    "mutmut",
    "black",
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.setuptools]
include-package-data = true